Refactor: 관리자 페이지 연결
- 페이지 이동 HomeController로 이동
